Move interface to internals
