+++ src/alley/wp/alleyvate/feature/class-user-enumeration-restrictions.php
@@ -12,7 +12,7 @@
 
 namespace Alley\WP\Alleyvate\Feature;
 
-use Alley\WP\Alleyvate\Feature;
+use Alley\WP\Alleyvate\Internals\Feature;
 use Stringable;
 use WP_Error;
 use WP_HTTP_Response;
